fix(core): Refactor project structure to resolve import errors
This commit fundamentally refactors the project structure to follow standard Python application layout, resolving critical runtime `ModuleNotFoundError` issues.

- **Adds `main.py` Entry Point:** A `main.py` file is introduced in the root directory to serve as the single, standard entry point for the application.
- **Centralizes Execution Logic:** The bot's startup logic is moved from `src/telegram_bot.py` into `main.py`.
- **Fixes All Imports:** All relative imports within the `src` package (e.g., `from .` or `from ..`) have been converted to absolute imports (e.g., `from src. ...`). This ensures that modules can be correctly resolved regardless of how the script is run.
- **Updates Documentation:** The `README.md` is updated to reflect the new, correct execution command (`python main.py`).
- **Ensures Stability:** The full test suite (26 tests) passes, confirming that the refactoring was successful and introduced no regressions.
