+++ README.md
@@ -26,7 +26,7 @@
 ├── .env.example          # ملف مثال لمتغيرات البيئة
 ├── populate_data.py      # سكربت لجلب البيانات التاريخية (لأغراض الاختبار المستقبلي)
 ├── requirements.txt      # الاعتماديات الخاصة ببايثون
-├── telegram_bot.py       # نقطة الدخول الرئيسية لتشغيل بوت تليجرام
+├── main.py               # نقطة الدخول الرئيسية لتشغيل البوت
 └── README.md
 ```
 
@@ -62,7 +62,7 @@
 1.  **شغل البوت:**
     تأكد من أنك في البيئة الافتراضية وأن ملف `.env` الخاص بك صحيح، ثم قم بتشغيل الأمر التالي:
     ```bash
-    python src/telegram_bot.py
+    python main.py
     ```
 
 2.  **تفاعل مع البوت على تليجرام:**
